requirements.txt
streamlit
watchdog
openai
python-dotenv
langchain
PyPDF2
qdrant-client
pycryptodome
tiktoken
youtube-transcript-api
pytube